Rename faces file as types
